More API changes

+++ tcpinterface/tcp.h
@@ -3,7 +3,7 @@
 int tcpserverinit(const std::string&);
 int tcpclientconnect(const std::string&, const std::string&);
 int tcpserverconnect(int, std::string&, std::string&); //blocking
-int tcprecv(int, field&);
-int tcpsend(int, field&);
+long int tcprecv(int, field&);
+long int tcpsend(int, field&);
 int tcpclose(int);
